fix circular deps in react-core
