+++ packages/react-core/src/components/ApplicationLauncher/ApplicationLauncherSeparator.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { DropdownSeparator, SeparatorProps } from '../Dropdown';
+import { DropdownSeparator, SeparatorProps } from '../Dropdown/DropdownSeparator';
 
 export const ApplicationLauncherSeparator: React.FunctionComponent<SeparatorProps> = ({
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
