+++ packages/react-core/src/components/Tabs/TabsContext.ts
@@ -1,0 +1,12 @@
+import * as React from 'react';
+
+export interface TabsContextProps {
+  variant: 'default' | 'light300';
+}
+
+const TabsContext = React.createContext<TabsContextProps>({
+  variant: 'default'
+});
+
+export const TabsContextProvider = TabsContext.Provider;
+export const TabsContextConsumer = TabsContext.Consumer;
