+++ packages/react-core/src/components/Tabs/index.ts
@@ -1,5 +1,6 @@
 export * from './Tab';
 export * from './Tabs';
 export * from './TabContent';
+export * from './TabsContext';
 export * from './TabTitleText';
 export * from './TabTitleIcon';
